Improve visibility of potential labels to add.
