+++ components/Labels/FilterLabelBox.module.scss
@@ -9,7 +9,7 @@
   flex-direction: column;
 
   width: 600px;
-  max-height: 90vh;
+  height: 80vh;
 
   animation-name: fade-in;
   animation-duration: 0.5s;
