+++ components/Labels/ManageLabelBox.module.scss
@@ -8,8 +8,8 @@
   display: flex;
   flex-direction: column;
 
-  width: min(90vw, 400px);
-  max-height: 70vh;
+  width: min(90vw, 600px);
+  height: 70vh;
 
   animation-name: fade-in;
   animation-duration: 0.5s;
@@ -32,16 +32,14 @@
   color: $equinor_PROMINENT;
 }
 
-.labelSection {
+.addSection {
   display: flex;
-  margin-top: 30px;
-  flex-wrap: wrap;
-  overflow: auto;
+  flex-direction: column;
 }
 
-.containerListExistingLabels {
-  position: absolute;
-  top: 37px;
-  z-index: 1000;
-  background-color: white;
+.labelSection {
+  display: flex;
+  flex-wrap: wrap;
+  overflow: auto;
+  margin-top: 30px;
 }
