+++ components/Labels/ManageLabelBox.tsx
@@ -1,12 +1,24 @@
 import React, { useEffect, useRef, useState } from "react";
-import { Button, Chip, Icon, Input, Scrim } from "@equinor/eds-core-react";
+import {
+  Button,
+  Chip,
+  Icon,
+  Input,
+  Scrim,
+  Search,
+} from "@equinor/eds-core-react";
 import styles from "./ManageLabelBox.module.scss";
 import { close } from "@equinor/eds-icons";
-import { addLabelToProcess, removeLabelFromProcess } from "services/labelsApi";
-import { useMutation, useQueryClient } from "react-query";
+import {
+  addLabelToProcess,
+  getLabels,
+  removeLabelFromProcess,
+} from "services/labelsApi";
+import { useMutation, useQuery, useQueryClient } from "react-query";
 import { vsmProject } from "interfaces/VsmProject";
 import { processLabel } from "interfaces/processLabel";
-import AddLabelInput from "./AddLabelInput";
+import { debounce } from "utils/debounce";
+import { unknownErrorToString } from "utils/isError";
 
 export default function ManageLabelBox(props: {
   isVisible: boolean;
@@ -20,7 +32,6 @@ export default function ManageLabelBox(props: {
       <div className={styles.box} onWheel={(e) => e.stopPropagation()}>
         <TopSection handleClose={props.handleClose} />
         <AddSection process={props.process} />
-        <LabelSection process={props.process} />
       </div>
     </Scrim>
   );
@@ -38,7 +49,27 @@ function TopSection(props: { handleClose: () => void }): JSX.Element {
 }
 
 function AddSection(props: { process: vsmProject }): JSX.Element {
+  const [term, setTerm] = useState("");
+  const [debouncedTerm, setDebouncedTerm] = useState("");
   const queryClient = useQueryClient();
+
+  const { data: labels, error } = useQuery(["labels", debouncedTerm], () =>
+    getLabels(debouncedTerm)
+  );
+
+  const handleChange = (event) => {
+    setTerm(event.target.value);
+  };
+
+  useEffect(() => {
+    debounce(
+      () => {
+        setDebouncedTerm(term);
+      },
+      200,
+      "LabelSearchQuery"
+    );
+  }, [term]);
 
   const addLabelMutation = useMutation(
     (payload: {
@@ -50,30 +81,18 @@ function AddSection(props: { process: vsmProject }): JSX.Element {
     { onSettled: () => queryClient.invalidateQueries() }
   );
 
-  const addLabel = (label: processLabel | { text: string }) => {
-    addLabelMutation.mutate({
-      processID: props.process.vsmProjectID,
-      label,
-    });
-  };
-
-  const handleSelectTerm = (term: string) => {
-    const trimmedTerm = term.trim();
-    if (!trimmedTerm) {
+  const handleSelect = (item: string) => {
+    const trimmedItem = item.trim();
+    if (!trimmedItem) {
       return null;
     }
-    addLabel({ text: trimmedTerm });
+    addLabelMutation.mutate({
+      processID: props.process.vsmProjectID,
+      label: { text: trimmedItem },
+    });
+    setTerm("");
   };
 
-  return (
-    <>
-      <AddLabelInput handleSelectTerm={handleSelectTerm} />
-    </>
-  );
-}
-
-function LabelSection(props: { process: vsmProject }): JSX.Element {
-  const queryClient = useQueryClient();
   const removeLabelMutation = useMutation(
     (payload: { processID: number; labelID: number }) =>
       removeLabelFromProcess(payload.processID, payload.labelID),
@@ -83,24 +102,57 @@ function LabelSection(props: { process: vsmProject }): JSX.Element {
   );
 
   return (
-    <div className={styles.labelSection}>
-      {props.process.labels.map((label) => (
-        <Chip
-          key={label.id}
-          onDelete={() =>
-            removeLabelMutation.mutate({
-              processID: props.process.vsmProjectID,
-              labelID: label.id,
-            })
+    <>
+      <Search
+        aria-label="search"
+        id="searchProjects"
+        placeholder="Search labels"
+        className={styles.searchField}
+        onChange={handleChange}
+        autoFocus
+      />
+      {error && <p>{unknownErrorToString(error)}</p>}
+      <div className={styles.labelSection}>
+        {props.process.labels?.map((label) => (
+          <Chip
+            key={label.id}
+            onDelete={() =>
+              removeLabelMutation.mutate({
+                processID: props.process.vsmProjectID,
+                labelID: label.id,
+              })
+            }
+            style={{
+              marginRight: "10px",
+              marginBottom: "10px",
+            }}
+            variant="active"
+          >
+            {label.text}
+          </Chip>
+        ))}
+
+        {labels?.map(function (label) {
+          if (
+            !props.process.labels.some(
+              (existingLabel) => existingLabel.text == label.text
+            )
+          ) {
+            return (
+              <Chip
+                key={label.id}
+                onClick={() => handleSelect(label.text)}
+                style={{
+                  marginRight: "10px",
+                  marginBottom: "10px",
+                }}
+              >
+                {label.text}
+              </Chip>
+            );
           }
-          style={{
-            marginRight: "10px",
-            marginBottom: "10px",
-          }}
-        >
-          {label.text}
-        </Chip>
-      ))}
-    </div>
+        })}
+      </div>
+    </>
   );
 }
